erl_docgen: Fix codeinclude
The regexp for finding <codeinclude> did not take into
consideration that elements in xml do not need to have
a closing tag. We shouldn't really try parsing xml with
regexp, but we try to make the best of the situation.

This commit also removed codeinclude from the dtd so that
at least in the future if any other codeinclude falls through
the regexp matching the xmllint will catch it.

Closes #5662