Improve register usage
Don't refetch a BEAM register whose contents already happen to be in a
CPU register (left there from the previous instruction).

This optimization makes most difference for the x86_64 JIT, where it
is applied about 57,000 times in the Erlang/OTP code base.

The AArch64 JIT keeps the 6 lowest-numbered X registers in CPU
registers, but the optimization is still applied about 10,000 times in
the Erlang/OTP code base.

Co-authored-by: John Högberg <john@erlang.org>