Propagate ranges for arithmetic operations
The main type analysis and optimizer sub pass iterates until a
fixpoint is reached. To ensure convergence, this sub pass is limited
in how it calculates ranges. It is only allowed to calculate a range
for operations that retain or shrink the ranges of their
operands. Examples of such operations are `band`, `bsr`, and `rem`. It
is not possible to calculate a range for operations such as `+' and
`bsl` that expand the ranges of their arguments.

This commit adds the `opt_range` sub pass that is run a single time
after type analysis pass. It will start from the types found in
annotations and propagate them forward through arithmetic instructions.

Here is an example:

    foo(N0) ->
        N1 = N0 band 3,
        N = N1 + 1,        % N1 is in 0..3
        element(N,         % N is in 1..4
            {zero,one,two,three}).

The range for `N1` was found by the main type pass. The range for
`N` was found by `opt_ranges`. Given a range for `N`, the JIT
will omit all range and type test for this call to `element/2`.