erts: Fix table info segfault during crash dump
When creating a crash dump there is no process available to
allocate terms to the heap. The new write_concurrency option
needs a heap to allocate a tuple on, so it would segfault
in the old code.

So the table_info function now takes a factory that we can use
to allocate all needed term in.