Update guidance on spec "when" for consistentcy
*For now*, standardize on the simple substitution
interpretation of `::` in constraints. Bounded
quantification could always be added later with
dedicated syntax (`<:` or `subtype_of`, etc.).

Previously, guidance for `when` in specs was that
tools could *decide* whether to implement
bounded quantificaiton or not.

That's a dangerous path in a world where
we want users and OTP team to be able to
write specs *once* and have tooling be interoperable.

So if we're going for consistency, then we have a choice
of whether to say `when` in specs is for bounded
quantification or simple substitution.

- Substitution interpretation: `-spec id_tup(X) -> X :: tuple()`
is equivalent to `-spec id_tup(tuple()) -> tuple().`
- Bounded quantification interpretation:
`-spec id_tuple(X) -> X :: tuple()` is generic and bounds
`X` to be a `tuple()`. This is the kind of thing that might
be written in made-up explicit syntax as
`-spec id_tuple<T extends tuple()>(T) -> T`.
`id_tuple(an_atom)` would lead to a type error and
`id_tuple({a, b})` would have return type `{a, b}`.

There are both pragmatic and theoretical reasons in favor of
the substitution interpretation.

**Reasons in Favor of The Substitution Interpretation**

- The substitution interpretation **reflects Erlang reality**.
   - Dialyzer, Gradualizer, and mystery-type-checker all use
   the substitution interpretation.
   - Most OTP specs don't make much sense on the bounded quanitification
   interpretation. `:: term()` appears over 2000 times in this repo,
   but why would someone bound something to be a subtype of `term()`?
- The substitution interpretation makes things easier for implementers,
including contributors to this repo. Dialyzer currently (iuc) doesn't
handle generics as generics at all. It sounds easier to add unbounded
genericity first rather than committing to the more complex bounded
quantification.
- The substitution interpretation is more syntactically consistent.
In `-type ty() :: atom()`, for example, the `::` means "is", not
"some type that is a subtype of atom()".
- I sometimes see Erlang devs confused about the difference between
type variables and `term()` ([example in
Q&A](https://youtu.be/_u1NDuFsW2A?t=1528)). Bounded quantification is yet another
rung of complexity for users.
- Bounded quantification is a complex feature that ideally would be
incorporated into the language in a holistic way:
    - For example, it would be an odd barrier to composability if
    specs could express bounds but type aliases and opaques could not.
    Such an asymmetry would prevent seemingly trivial refactors such
    as extracting out parts of a long spec into type aliases.
    - To take another example, if bounds really are part of types,
    then error reporting for static analysis ideally would include
    bounds as well. It takes some work to do this in a way that
    is both informative and not befuddling.
    - Variance interacts with bounded quantification,
    but the variance story is still being written.
    [@jhogberg is investigating variance annotations for aliases
    and
    opaques](https://github.com/erlang/otp/pull/5223#issuecomment-924917745).

**Note on Backwards-Compatibility**

This is a minor change in official guidance.
This change in guidance won't break anything as far as I can tell,
so I don't think we tarnish Erlang's great reputation as a stable, backwards-compatible
language. And we improve the situation going forward,
since the previous guidance said implementators could make their
own choices and now we're standardizing and preventing the type-writing community
from splitting in the future.