Code cleanup
Re-use and restructure type checking functions in prim_inet.

Fix a bug in gen_sctp:connect that did not handle port number
as a service name correctly.

Align gen_sctp:connectx_init more with gen_udp:send that is
one of the few functions to handle different kinds of addresses,
for SockAddr style addresses.

Drop the combination of SockAddrs and Timeout since it is
not meaningful.  The Timeout is only used for address lookup
and SockAddrs are already looked up.

Fix the type spec for connecx_init to show that Port
can be a Service.

Clean up the code that enforces the same port number in
all addresses to be more efficient.

Write a dedicated function loop to look up addresses instead
of using throw() from a lists:map/2.