Make gen_server doc use type specs
Since there now are type specs I have moved the explanation of
of the type values to the type.  Previously the it had to be
under a selected function and then other functions referred
to that function.  Now all functions refer to the type.

For example; under the type start_ret() the possible return
values for all start functions (start/3,4, start_link/3,4
and start_monitor/3,4) are described and all start functions
refer to start_ret().  Previously the return values were
described under start_link/3,4 and the other start functions
referred to it.

Now you often need to follow a link to find the explanation
of a value, but it is described together with the auto
generated type spec so it is less likely that the description
starts diverting from the type spec, and the description
can refer to annotation variables in the spec.

In short, information duplication has been reduced by
moving the information to a type, but that requires
the reader to follow a link to find the information
from a function description.