Rework diameter_{traffic,tls,transport}_SUITE
Make them runnable without common_test, by way of run/0, which is
convenient for running suites easily from the shell; for example:

  $ env ERL_CRASH_DUMP_BYTES=0 erl -noinput -pa ebin -pa test -s diameter_traffic_SUITE run -s init stop

Just crashing also provides better diagnostics than common_test often
provides in the shell; for example:

  diameter_traffic_SUITE:init_per_group failed
  Reason: {badmatch,undefined}

  rand:uniform_s failed on line 334
  Reason: {function_clause,[{rand,...},{...}|...]}

Having to look in html pages to find details of failures like this isn't
ideal.

Testcases were also dependent on one another, so one failure could lead
to several. Now run everthing that needs to be done in sequence, and
different tests within a sequence in parallel.

Multiple simultaneous connections in the transport suite are often
refused on macOS, so be satisfied with one.