Rework remaining diameter testsuites
As in the previous two commit: remove common_test as a requirement and
make all testcases independent. Add a make target for running any suite
without common test; for example, to run all testcases in
diameter_dpr_SUITE:

  make _dpr

Similar to the target that runs suites with common_test, just add an
underscore prefix.