Split TypEr core out of typer module.
We (@pablocostass, @maco, and myself) are building rebar3_typer and, as part of our efforts to integrate TypEr into rebar3, we found that typer (the module) didn't expose any interface that allowed us to use it from another Erlang module. That's why we decided to move some pieces of it into our project. In particular: we created typer_core with just the core functionality, detached from the parts that parse the command-line arguments and options.
Eventually, we came up with AdRoll/rebar3_typer#15, and we thought it would be great if we could contribute this back to OTP itself, so… for future versions of the plugin, we don't need to use our own version of typer and we can just use what comes with OTP.
As a bonus, we added a new mode: annotate_in_place. It's like annotate but it replaces the original files instead of creating new ones.

In our project, we also added several tests for typer_core. I didn't move them here since I don't know if they match the test structure of this project. If you want to check them out and let us know if you like them, we can migrate them here of course.