epmd: don't fail if non-critical interfaces fail to bind
When specifying bind address(es) via ERL_EPMD_ADDRESS or the -address
option, epmd will bind to the IPv4/IPv6 loopback intefaces in addition
to the user-supplied address(es).

However, it is possible that one of the loopback interfaces does not
exist.  For example, if the host system has ipv6 disabled via the
disable_ipv6 sysctl, there will be no loopback interface "::1", and
attmepting to bind to "::1" will fail.

Under these circumstances, log the failure but do not consider the
failure to bind to the loopback interface as fatal.

If any of the user-supplied addresses fail to bind, however, that is
still considered fatal and epmd will terminate.