[erts] Expose parent process via process_info(_, parent)
Parent process information is available in the VM process structure,
but is not exposed via process_info. Sometimes it is very convenient
to know who spawned processes that are not supporting any OTP
behaviours (where this information is stored in a process dictionary).

Parent process is returned only for locally started processes. Remotely
started process, or 'init' process that has no parent, return 'undefined'.