logger: Don't use original meta data in internal log
logger_backend emits internal debug logs if a handler or a filter
fails. This is done using the ?LOG_INTERNAL macro. Prior to this
commit, these logs would get the meta data from the original failing
log event. This commit instead uses the actual location of the
internal debug log.

The original meta data is still included as a part of the log report,
so no data is lost.