erl_tar: Slightly optimize extraction from tar files
Running eprof when extracting tar files revealed that the top CPU
time consumers in the erl_tar process itself (that is, not processes
doing file IO) were checksum calculation and parsing of octal numbers.

This commit optimizes those functions. It turned out that the
checksum is calculated in two different ways, although the second
signed checksum is only neeeded if the first checksum is incorrect
(when extracting very old tar files).