Add appup for diameter 2.2.5
The parent commit modified modules in diameter, but since there have
been no other changes in diameter there's no appup entry that would let
the commit pass unnoticed: our release machinery forces a version bump
for OTP 24.3, and that requires an appup entry for upgrade to be
possible (and for our system upgrade suite to pass), even though none of
the modified modules are in the app file. This all seems a little
questionable, but add an appup entry for the noop upgrade to pacify the
machinery.

The changes in diameter are to info/debug modules, not released code.
That is, despite the version bump, there's no change to released modules
in diameter.