Decode/encode Address types other than IP and IP6
RFC 6733 defines the Diameter Address type as follows.

  4.3.1.  Common Derived AVP Data Formats

     Address

        The Address format is derived from the OctetString Basic AVP
        Format.  It is a discriminated union representing, for example, a
        32-bit (IPv4) [RFC0791] or 128-bit (IPv6) [RFC4291] address, most
        significant octet first.  The first two octets of the Address AVP
        represent the AddressType, which contains an Address Family,
        defined in [IANAADFAM].  The AddressType is used to discriminate
        the content and format of the remaining octets.

IP and IP6 are AddressType 1 and 2 respectively, but there are many more
defined in IANAADFAM, here:

  http://www.iana.org/assignments/address-family-numbers/address-family-numbers.xhtml

None of these additional values were decoded/encoded in module
diameter_types. All were treated as errors.

The definition of other address families is spread across many different
documents, not all of which are easily accessible or even identified by
IANAADFAM, and the list can change. Given that the current decode has
been in place for over 10 years now (ie. this has clearly not been a big
problem), only provide a minimal decode of non-IP address families,
splitting them into a 2-tuple of unreserved 16-bit family and remaining
binary with no family-specific decode. Encode maps the 2-tuples back to
binary, and accepts address families 1 and 2 as well.

Note that even currently unassigned address families are decoded: only
the reserved values (0 and 65535) are treated as errors.

Decoding the binary to a 2-tuple at all is debatable since there's
hardly anything to gain over simply treating Address like OctetString,
with only validation that it contains an unreserved 2-octet AddressType,
but choose the tuple for consistency with the IP address decodes. It's
still possible to add family-specific decodes in the future with a
decode option, similar to the existing {string_decode, boolean()}.