Return empty history when failing to create a log
This fixes an infinite loop with group:up_stack/1 that could be
triggered when commandline history is enabled, but it wasn't possible to
create the log file. In this case, an `ok` atom was being stored as the
history which would confuse the history search.

If you pressed CTRL-R and started typing, an infinite loop would be
started that would eventually terminate when out of memory.

To reproduce:

```sh
$ export ERL_AFLAGS="-kernel shell_history enabled -kernel shell_history_path '\"/proc\"'"
$ erl

Type CTRL-R and then press any key
```