observer: Call original event handlers for menu commands
This is important for macOS where by default we get the "App menu" with:

- Services
- Hide <app>
- Hide Others
- Show All
- Quit <app>

Before this patch, "Hide <app>", "Hide Others", and "Show All" are not
doing anything. This is what we get when clicking "Hide <app>":

    observer_sys_wx:238: Unhandled event {wx,5250,
                                          {wx_ref,35,wxFrame,[]},
                                          [],
                                          {wxCommand,command_menu_selected,[],-1,
                                           0}}

After this patch, the original (system-level) event handlers are fired
and on the Erlang side we can simply ignore these events.