[megaco|test] Tweaked the action test suite
Adjusted to make use of the try-tc utility.
Also add some more (debug) printouts during node start.
Ensure nodes are (explicitly) stopped at the end of tc.
Add a wrapper to ping to handle if the net_adm:ping takes
a long time (which it can). Skip in those cases.