compiler: Improve deterministic builds
Improves build determinism for non-native code by:
- Making codegen respect the +deterministic option
- Adding an ERL_DETERMINISTIC env var that controls OTP determinisim

This addresses absolute paths being included in generated .erl files
and compiled .beam files that resulted in builds from different source
directories generating different artefacts (which is a component of the
issue in #4482).

OTP itself can now be built in deterministic mode by setting
ERL_DETERMINISTIC=yes in the environment. I think it would make sense to
make this the default at some stage, but I've put the change behind an
option for now in an attempt to decouple making deterministic OTP build
possible from making them the default.

Having +deterministic set also results in the compiler options being
removed from the module info for modules where this options was used.
This may have other implications for users of OTP.

For tests themselves, +determinism is disabled, since many test cases
depend on accessing the test module's compilation options or other
features not available in deterministic mode, in order to configure
themselves. For tests of the determinism feature, +determinism is
explicitly passed to the compiler within the relevant test cases.