wx: Add MacGetCommonMenuBar and MacSetCommonMenuBar
This is especially useful for apps that manage multiple wxFrames where
we can close the last window but want the application to keep running.
Before this patch we'd be stuck with the default menu bar, created in
`wxe_impl.cpp`, that we cannot interact with. After this patch we can,
which means we can add new menu items, subscribe to events, etc.

Example:

    wx:new(), MB = wxMenuBar:macGetCommonMenuBar(), M = wxMenuBar:oSXGetAppleMenu(MB), wxMenu:setTitle(M, "Test").