beam_call_types: Correct optimization of setelement/3
2ea04617d18bc3 introduced a bug in the handling of calls to
`setelement/3`. Consider this module:

    -module(bug).
    -export([bug/0]).

    bug() ->
        set(2, {a,b}),
        set(3, {a,b}).

    set(Pos, Tuple) ->
        setelement(Pos, Tuple, something).

The second call to `set/2` will fail because the position is `3` is
beyond the end of the tuple. Because of that failed call, the compiler
would incorrectly assume that the call to `setelement/3` in `set/2`
would always fail, and there would be a very confusing exception:

    1> bug:bug().
    ** exception error: no true branch found when evaluating an if expression
         in function  bug:set/2 (bug.erl, line 8)

This commit fixes the bug. With this bug fix, the exception will look
like:

    4> bug:bug().
    ** exception error: bad argument
         in function  setelement/3
            called as setelement(3,{a,b},something)
            *** argument 1: out of range
         in call from bug:set/2 (bug.erl, line 9)