logger: Fix repeated filesync shutdown timeout
The commit 06cacbc (aka PR #4811) changed where the timer_server
is located in the supervisor tree of Kernel. This change caused
the timer_server to be terminated before logger, which meant that
the calls to timer:cancel from any logger backend would timeout.

We solve this by instead using erlang:start_timer for the repeated
filesync timer.

Closes #5780