[observer] What to do with "bad" options
Include "bad" options in the list, but not with a "value"
(since we did not get any). Instead with something that
will explain "something": A string; either "-",
"Not Supported" or "error: ..."
We can see what is missing and why (its better then nothing).
Also, this is approx like what is done for socket.

OTP-18012