erts: Shrink ErlFunThing structure
The sum of `arity` and `num_free` cannot exceed 255, so there's no
point using a word for each.

Note that we will now badarg when passing an arity above 255 to
`erlang:make_fun/3`, instead of creating a fun that can never be
called and had iffy behavior when sent over the distribution (as
the arity got truncated down to 8 bits. wat), but it ought to
be okay as its spec has always said the valid range is 0..255.