diffable: Allow input from external runtime
When `diffable` is used to evaluate the impact of a compiler
optimization or checking that a compiler refactoring doesn't change
the result in any significant way, the changes to the compiler source
code produce many differences in the output from `diffable`. The sheer
number of changes in line number directives in the output from, for
example, a reformatting which inserts a line early in a module makes
it hard to notice significant changes in the output.

Prior to this patch `diffable` finds the source code to the
applications it compiles by using `code:root_dir/0` and
`code:lib_dir/1`, that is, it assumes that the source code to compile
is the source for the running system. This patch adds a new command
line flag to `diffable`. The flag, `--erltop <path>`, makes diffable
look for the source code for the applications to compile in a
different tree.

The intended use case is as follows, you have a baseline and a
development tree, located in directories $BASE and $DEV
respectively:

```
$ PATH=$BASE/bin:$PATH $BASE/scripts/diffable --asm \
      --deterministic /tmp/baseline
```

Now run `diffable` using the development compiler, but compile the
source code from the baseline:

```
$ PATH=$DEV/bin:$PATH $DEV/scripts/diffable --erltop $BASE --asm \
      --deterministic /tmp/dev
```