erts: Fix localtime_r summer/winter-time change bug
According to POSIX, localtime_r does not have to update the process
internal tz data when called. So if time went from winter to
summertime erts is running localtime_r may return an incorrect time.

So we need to make sure to call tzset before each call to localtime_r.
localtime is guaranteed to update the tz information, so no need to
call it when using localtime.