erts: Fix a race in persistent_term:get/1,2
If a certain key did not exist in the hash table, we'd abort the
lookup as soon as we noticed that and return the current index.

When we then called get_bucket, another key-value pair could have
been inserted at that index, making us return the value of that
pair instead.