Fix incorrect optimization of setelement/3
`beam_call_types` would assume that a call to `setelement/3`
would always fail for a tuple with known exact size and a position
that could be both inside and outside the tuple. That would make
the call to `setelement/3` tail-recursive and all code following
the call would be ignored.

This bug was introduced in 2ea04617d18b. It has already been fixed on
the `master` branch (in #5820) with an updated test case. To avoid
cumbersome merge conflicts, this commit don't include any test case.