erts: Change erlang:display_string to output to stdout
Historically, display_string has been used to output to stderr
when the standard_error io device is not yet available. However
all usage of it for that purpose has been removed/rewritten and
now it is used to write a string together with erlang:display.
So we change it to write to stdout (as erlang:display does).