Improve seeding
The seed function should not produce the same generator state
for any two user input seeds.  Restrict the user input and use
reversible (bijectional) mappings to produce the seed.

Suggest using multiply-and-shift for range capping in the
documentation and tests.  This will be revised in a later
commit since it is only a good idea if the high bits are good.