JIT: Optimize creation of binaries
In the JIT, improve the code generation for binary construction
for segments of fixed widths of no more than 64 bits and for
variable width segments that are a set to 0.

I have run the benchmark linked to from https://github.com/erlang/otp/issues/5639.

On my Intel iMac from 2017, the result without this commit was:

    == Testing with 1 MB ==
    fun base64:encode/1: 1000 iterations in 16264 ms: 61 it/sec
    fun base64:decode/1: 1000 iterations in 18597 ms: 53 it/sec

With this commit:

    == Testing with 1 MB ==
    fun base64:encode/1: 1000 iterations in 10955 ms: 91 it/sec
    fun base64:decode/1: 1000 iterations in 10629 ms: 94 it/sec

On my M1 MacBook Pro from 2020 the result before was:

    == Testing with 1 MB ==
    fun base64:encode/1: 1000 iterations in 12299 ms: 81 it/sec
    fun base64:decode/1: 1000 iterations in 15147 ms: 66 it/sec

After:

    == Testing with 1 MB ==
    fun base64:encode/1: 1000 iterations in 8550 ms: 116 it/sec
    fun base64:decode/1: 1000 iterations in 10066 ms: 99 it/sec