Fix type specs for SCTP get/set opts, revised
Revert the exported type names in gen_sctp to option() and
option_name().  We cannot change them since they are a part of the API.
So we should be backwards compatible and keep the names although
they are somewhat misleading.

Add gen_sctp:option_value() for the returned option values, and use that
in an also added inet:socket_optval().  Use the latter for the return
value of inet:getopts/2, which is the a correction and therefore
not backwards incompatible.

Separate the "Data Types" section in gen_sctp and inet into
"Exported data types" and "Internal data types", and place
the defined types in the right sub section.