erts: Improve internal hashing
CRC32-C has better distribution and collision qualities than Bob
Jenkins' classic hash function, and is much faster when we can use
dedicated processor instructions.

Granted, it's not _great_ and there's a lot of work remaining
before I'm happy with both the hash quality and the code itself,
but it's much better than it was before.