Correct the type spec for format_status/1
The bad type spec in gen_statem, gen_server and gen_event
stated that the return value of format_status/1
is exactly the input argument value, which is incorrect.

To fix this, a type format_status() was needed in each module
to be used both for the argument and the return value.  This type
is exported since it is essential for the format_status/1 function.

Also, for gen_statem, a type event_content() :: term()
was needed to facilitate documentation and to not lure any future
code update into incorrectly using a type value EventContent
multiple times in a bad way.